Give service users full s3 access
It's not clear what the exact bucket access should be for a bridgepf
service user so we will give the user AmazonS3FullAccess for now.

Note - Providing AmazonS3FullAccess basically nullifies the
stricter IAMBridgepfS3ManagedPolicy and IAMBridgepfLocalS3ManagedPolicy
policies.  I've decided to just add full access now and leave the cleanup
to a seperate change.