Fix deployment of vars to system parameter store
The deploy_infra script was exiting right after deployment of cf
template.  Now that we have other things to deploy we want it to
deploy multiple things.  This change de-couples normalizing vars
and deployments. Now we have a separate file for interpolating vars,
deploying CF template and deploying SSM params.  This will make it
easier to add on additional deployments if needed.