BRIDGE-2038: EB role for BridgePF app
The Bridge app uses AWS heroku access keys to access AWS resources.
The problem with this is that the heroku account is an admin account
and has too much privileges. For a more secure environment we want
to enforce a least privilege model of access.  The better approach
for that is to setup access with roles and instance profiles. This
change is to setup AWS EB role to allow EC2 instances to access
specific AWS resources (s3, ddb, rds, etc..).