BRIDGE-2057: Replace deprecated DB VPC security group.
"DB VPC security groups have been deprecated"[1] which means they
do not appear in the AWS console and cannot be attached to a DB
instance.  The recommendation is to replace it with an EC2 security
group.

The TCP inbound setting is to allow EC2 instances launch by EB to access
the DB.  The CidrIp inbound setting is to allow connections to the DB from
inside the Fred huch VPN.

[1] https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Overview.RDSSecurityGroups.html