BRIDGE-2057: Replace safelist with wide open policy
BridgePF tests requires travis sudo and a direct connection to the
dev environment (rds instance). The travis safelist[1] required for sudo
is quite big so the strategy is to apply a wide open DB policy to the Dev RDS
instance instead of adding the entire safelist of IPs to the policy.
This wide open policy should only be applied to the Dev RDS instance.

[1] https://docs.travis-ci.com/user/ip-addresses/