Let CF name the user signed consents bucket
Giving CF Buckets specific names makes it more difficult to
migrate bridgepf to a separate aws account.  The reason is that
buckets names cannot be duplicated accross aws accounts.  If we
give it a specific name in one account and attempt to run the CF
script again in another account it will fail because CF will
try to give it the new bucket the same name.  The better approach
and one that I've found to be more flexible is to let CF manage
resource names.  We create a bucket letting CF name it then
set parameters in bridge-server.conf with the new bucket name.
We can remove the deprecated bucket after bridgepf app is updated
to reference the new bucket.