increase prod hosts size and type
BP Lab redrives is currently spiking production and putting it into an unhealthy state. We're increasing prod instance types from t2.small to t3.medium, and setting the min hosts to 6 so we always have 6 hosts.

We will decrease the number and type of hosts after the redrive is complete (~mid October) and revisit auto-scaling.